+++ app/Http/Controllers/IndicatorController.php
@@ -7,24 +7,28 @@ use App\Models\Master;
 
 class IndicatorController extends Controller
 {
-    public function home()
+    public function home($tahun)
     {
-        $master = Master::where('jenis','Indikator')->get();
-        $aspek = Master::where('jenis','Aspek')->select('id','urutan')->get();
+        $master = Master::where('jenis','Indikator')->where('tahun',$tahun)->get();
+        $aspek = Master::where('jenis','Aspek')->where('tahun',$tahun)->select('id','urutan')->get();
         $data = [
             'master'=>$master,
-            'aspek'=>$aspek
+            'aspek'=>$aspek,
+            'tahun'=>$tahun
         ];
         return view('master.indikator',$data);
     }
-    public function add(Request $req)
+    public function add(Request $req,$tahun)
     {
-        $count = count(Master::where('jenis','Indikator')->get())+1;
+        $count = count(Master::where('jenis','Indikator')->where('tahun',$tahun)->get())+1;
         $master = new Master();
         $master->urutan = $count;
         $master->nama = $req->nama;
+        $master->keterangan = $req->keterangan;
         $master->jenis = 'Indikator';
         $master->bobot = $req->bobot;
+        $master->skor = 0;
+        $master->tahun = $tahun;
         $master->catatan = $req->catatan;
         $master->analisis = $req->analisis;
         $master->rekomendasi = $req->rekomendasi;
@@ -35,7 +39,8 @@ class IndicatorController extends Controller
     public function edit($id)
     {
         $master = Master::find($id);
-        $aspek = Master::where('jenis','Aspek')->select('id','urutan')->get();
+        $tahun = $master->tahun;
+        $aspek = Master::where('jenis','Aspek')->where('tahun',$tahun)->select('id','urutan')->get();
         $data = [
             'master'=>$master,
             'aspek'=>$aspek
@@ -46,18 +51,25 @@ class IndicatorController extends Controller
     {
         $master = Master::find($id);
         $master->nama = $req->nama;
+        $master->keterangan = $req->keterangan;
         $master->bobot = $req->bobot;
         $master->catatan = $req->catatan;
         $master->analisis = $req->analisis;
         $master->rekomendasi = $req->rekomendasi;
         $master->id_parent = $req->id_parent;
         $master->save();
-        return redirect('/indikator')->with('success','Berhasil');
+        $tahun = $master->tahun;
+        return redirect('/indikator/'.$tahun)->with('success','Berhasil');
     }
     public function delete($id)
     {
         $master = Master::find($id);
-        $master->delete();
-        return redirect()->back()->with('warning','Berhasil');
+        $child = Master::where('jenis','Parameter')->where('id_parent',$master->id)->get();
+        if (count($child) == 0) {
+            $master->delete();
+            return redirect()->back()->with(['warning' => 'Pesan Berhasil']);
+        } else {
+            return redirect()->back()->with(['info' => 'Gagal Hapus']);
+        }
     }
 }
